requirements.txt
feedparser
discord.py
html2text